refactor: [#66] NumberBaseballController 를 분리해서 ProgramController, GameController 로 만들기
NumberBaseballController 클래스가 혼자서 너무 많은 역할을 맡고 있는 듯 해서
Program 과 Game 두 가지 역할로 분리한다.

관련 일감 링크
[NumberBaseballController 를 분리해서 ProgramController, GameController 로 만들기](https://github.com/OptimistLabyrinth/java-baseball-precourse/issues/66)
