+++ src/main/java/baseball/constant/NumberBaseballGameConstant.java
@@ -1,11 +1,8 @@
 package baseball.constant;
 
-public class NumberBaseballConstant {
+public class NumberBaseballGameConstant {
     public static final Integer LENGTH_OF_TARGET_NUMBER = 3;
 
     public static final Integer MINIMUM_OF_EACH_DIGIT = 1;
     public static final Integer MAXIMUM_OF_EACH_DIGIT = 9;
-
-    public static final Integer GAME_RESTART = 1;
-    public static final Integer PROGRAM_EXIT = 2;
 }
