+++ src/main/java/baseball/constant/NumberBaseballProgramConstant.java
@@ -1,0 +1,6 @@
+package baseball.constant;
+
+public class NumberBaseballProgramConstant {
+    public static final Integer GAME_RESTART = 1;
+    public static final Integer PROGRAM_EXIT = 2;
+}
