+++ src/main/java/baseball/model/RandomNumberGenerator.java
@@ -1,6 +1,6 @@
 package baseball.model;
 
-import baseball.constant.NumberBaseballConstant;
+import baseball.constant.NumberBaseballGameConstant;
 import camp.nextstep.edu.missionutils.Randoms;
 
 import java.util.HashSet;
@@ -16,7 +16,7 @@ public class RandomNumberGenerator {
     }
 
     public String generate() {
-        while (randomNumberStringBuilder.length() < NumberBaseballConstant.LENGTH_OF_TARGET_NUMBER) {
+        while (randomNumberStringBuilder.length() < NumberBaseballGameConstant.LENGTH_OF_TARGET_NUMBER) {
             generateOneDigit();
         }
         return randomNumberStringBuilder.toString();
@@ -24,8 +24,8 @@ public class RandomNumberGenerator {
 
     private void generateOneDigit() {
         Integer randomNumber = Randoms.pickNumberInRange(
-                NumberBaseballConstant.MINIMUM_OF_EACH_DIGIT,
-                NumberBaseballConstant.MAXIMUM_OF_EACH_DIGIT
+                NumberBaseballGameConstant.MINIMUM_OF_EACH_DIGIT,
+                NumberBaseballGameConstant.MAXIMUM_OF_EACH_DIGIT
         );
         if (!numbers.contains(randomNumber)) {
             randomNumberStringBuilder.append(randomNumber);
