+++ src/main/java/baseball/model/inputvalidator/GameRestartInputValidator.java
@@ -1,12 +1,12 @@
 package baseball.model.inputvalidator;
 
-import baseball.constant.NumberBaseballConstant;
+import baseball.constant.NumberBaseballProgramConstant;
 
 public class GameRestartInputValidator implements InputValidator {
     @Override
     public Boolean validate(String input) {
         return input != null && (
-                input.equals(NumberBaseballConstant.GAME_RESTART.toString()) ||
-                        input.equals(NumberBaseballConstant.PROGRAM_EXIT.toString()));
+                input.equals(NumberBaseballProgramConstant.GAME_RESTART.toString()) ||
+                        input.equals(NumberBaseballProgramConstant.PROGRAM_EXIT.toString()));
     }
 }
